String Concatenation
Operators aren't just for numbers! When a + operator is used on two strings, it appends the right string to the left string.
This process of appending one string to another is called concatenation. Notice in the third example we had to make sure to include a space at the end of the first string. The computer will join the strings exactly, so we needed to make sure to include the space we wanted between the two strings.
Just like with regular math, we can combine, or chain, our operations to get a final result.